+++ server/src/main/java/org/opensearch/common/time/JavaDateFormatter.java
@@ -32,6 +32,7 @@
 
 package org.opensearch.common.time;
 
+import org.opensearch.common.util.FeatureFlags;
 import org.opensearch.core.common.Strings;
 
 import java.text.ParsePosition;
@@ -70,6 +71,7 @@ class JavaDateFormatter implements DateFormatter {
     private final DateTimeFormatter printer;
     private final List<DateTimeFormatter> parsers;
     private final JavaDateFormatter roundupParser;
+    private final Boolean canCacheLastParsedFormatter;
 
     /**
      * A round up formatter
@@ -78,8 +80,8 @@ class JavaDateFormatter implements DateFormatter {
      */
     static class RoundUpFormatter extends JavaDateFormatter {
 
-        RoundUpFormatter(String format, List<DateTimeFormatter> roundUpParsers) {
-            super(format, firstFrom(roundUpParsers), null, roundUpParsers);
+        RoundUpFormatter(String format, List<DateTimeFormatter> roundUpParsers, Boolean canCacheLastParsedFormatter) {
+            super(format, firstFrom(roundUpParsers), null, roundUpParsers, canCacheLastParsedFormatter);
         }
 
         private static DateTimeFormatter firstFrom(List<DateTimeFormatter> roundUpParsers) {
@@ -93,8 +95,12 @@ class JavaDateFormatter implements DateFormatter {
     }
 
     // named formatters use default roundUpParser
+    JavaDateFormatter(String format, DateTimeFormatter printer, Boolean canCacheLastParsedFormatter, DateTimeFormatter... parsers) {
+        this(format, printer, ROUND_UP_BASE_FIELDS, canCacheLastParsedFormatter, parsers);
+    }
+
     JavaDateFormatter(String format, DateTimeFormatter printer, DateTimeFormatter... parsers) {
-        this(format, printer, ROUND_UP_BASE_FIELDS, parsers);
+        this(format, printer, false, parsers);
     }
 
     private static final BiConsumer<DateTimeFormatterBuilder, DateTimeFormatter> ROUND_UP_BASE_FIELDS = (builder, parser) -> {
@@ -113,6 +119,7 @@ class JavaDateFormatter implements DateFormatter {
         String format,
         DateTimeFormatter printer,
         BiConsumer<DateTimeFormatterBuilder, DateTimeFormatter> roundupParserConsumer,
+        Boolean canCacheLastParsedFormatter,
         DateTimeFormatter... parsers
     ) {
         if (printer == null) {
@@ -128,6 +135,7 @@ class JavaDateFormatter implements DateFormatter {
         }
         this.printer = printer;
         this.format = format;
+        this.canCacheLastParsedFormatter = canCacheLastParsedFormatter;
 
         if (parsers.length == 0) {
             this.parsers = Collections.singletonList(printer);
@@ -135,7 +143,16 @@ class JavaDateFormatter implements DateFormatter {
             this.parsers = Arrays.asList(parsers);
         }
         List<DateTimeFormatter> roundUp = createRoundUpParser(format, roundupParserConsumer);
-        this.roundupParser = new RoundUpFormatter(format, roundUp);
+        this.roundupParser = new RoundUpFormatter(format, roundUp, canCacheLastParsedFormatter);
+    }
+
+    JavaDateFormatter(
+        String format,
+        DateTimeFormatter printer,
+        BiConsumer<DateTimeFormatterBuilder, DateTimeFormatter> roundupParserConsumer,
+        DateTimeFormatter... parsers
+    ) {
+        this(format, printer, roundupParserConsumer, false, parsers);
     }
 
     /**
@@ -164,7 +181,7 @@ class JavaDateFormatter implements DateFormatter {
         return null;
     }
 
-    public static DateFormatter combined(String input, List<DateFormatter> formatters) {
+    public static DateFormatter combined(String input, List<DateFormatter> formatters, Boolean canCacheLastParsedFormatter) {
         assert formatters.size() > 0;
 
         List<DateTimeFormatter> parsers = new ArrayList<>(formatters.size());
@@ -181,19 +198,21 @@ class JavaDateFormatter implements DateFormatter {
             roundUpParsers.addAll(javaDateFormatter.getRoundupParser().getParsers());
         }
 
-        return new JavaDateFormatter(input, printer, roundUpParsers, parsers);
+        return new JavaDateFormatter(input, printer, roundUpParsers, parsers, canCacheLastParsedFormatter);
     }
 
     private JavaDateFormatter(
         String format,
         DateTimeFormatter printer,
         List<DateTimeFormatter> roundUpParsers,
-        List<DateTimeFormatter> parsers
+        List<DateTimeFormatter> parsers,
+        Boolean canCacheLastParsedFormatter
     ) {
         this.format = format;
         this.printer = printer;
-        this.roundupParser = roundUpParsers != null ? new RoundUpFormatter(format, roundUpParsers) : null;
+        this.roundupParser = roundUpParsers != null ? new RoundUpFormatter(format, roundUpParsers, canCacheLastParsedFormatter) : null;
         this.parsers = parsers;
+        this.canCacheLastParsedFormatter = canCacheLastParsedFormatter;
     }
 
     JavaDateFormatter getRoundupParser() {
@@ -237,6 +256,12 @@ class JavaDateFormatter implements DateFormatter {
                 ParsePosition pos = new ParsePosition(0);
                 Object object = formatter.toFormat().parseObject(input, pos);
                 if (parsingSucceeded(object, input, pos)) {
+                    if (FeatureFlags.isEnabled(FeatureFlags.DATETIME_FORMATTER_CACHING)
+                        && canCacheLastParsedFormatter
+                        && formatter != parsers.get(0)) {
+                        parsers.remove(formatter);
+                        parsers.add(0, formatter);
+                    }
                     return (TemporalAccessor) object;
                 }
             }
@@ -260,7 +285,7 @@ class JavaDateFormatter implements DateFormatter {
             .stream()
             .map(p -> p.withZone(zoneId))
             .collect(Collectors.toList());
-        return new JavaDateFormatter(format, printer.withZone(zoneId), roundUpParsers, parsers);
+        return new JavaDateFormatter(format, printer.withZone(zoneId), roundUpParsers, parsers, this.canCacheLastParsedFormatter);
     }
 
     @Override
@@ -274,7 +299,7 @@ class JavaDateFormatter implements DateFormatter {
             .stream()
             .map(p -> p.withLocale(locale))
             .collect(Collectors.toList());
-        return new JavaDateFormatter(format, printer.withLocale(locale), roundUpParsers, parsers);
+        return new JavaDateFormatter(format, printer.withLocale(locale), roundUpParsers, parsers, this.canCacheLastParsedFormatter);
     }
 
     @Override